+++ packages/contracts/deploy/10_dao-registry.ts
@@ -9,6 +9,10 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
   const {deployer} = await getNamedAccounts();
   const managingDAOAddress = await getContractAddress('DAO', hre);
+  const ensSubdomainRegistrarAddress = await getContractAddress(
+    'ENSSubdomainRegistrar',
+    hre
+  );
 
   await deploy('DAORegistry', {
     from: deployer,
@@ -21,7 +25,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
       execute: {
         init: {
           methodName: 'initialize',
-          args: [managingDAOAddress],
+          args: [managingDAOAddress, ensSubdomainRegistrarAddress],
         },
       },
     },